Rename to pathpipeconftools
